Fixed unit tests and formatting
